requirements.txt
fastapi[all]>=0.111.0
uvicorn[standard]>=0.30.1
sqlalchemy>=2.0.30
alembic>=1.13.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
pydantic-settings>=2.3.4
Pillow>=10.3.0
mss>=9.0.1
pynput>=1.7.6
psutil>=6.0.0
requests>=2.32.3
langchain>=0.2.6
langchain-community>=0.2.6
bs4
